requirements.txt
web3==6.15.1
psycopg2-binary
python-dotenv==1.0.1
